requirements.txt
crewai==0.22.5
langchain-google-genai==1.0.1
python-dotenv==1.0.0
tavily-python==0.3.3
pydantic==2.5.0
flask==3.0.0